chore(deps): an advisory nobody is watching for is one nobody finds
The tree carried a live RUSTSEC vulnerability (crossbeam-epoch, invalid pointer
deref in the `Debug` impl for `Atomic`/`Shared`), a yanked `spin`, and five
further advisories. Nothing here had gone wrong yet — that is the point. Nothing
was watching, so they were found by a review that happened to be asked for, and
would otherwise have sat there.

Most of it was not a code problem. A plain `cargo update` cleared the
vulnerability, the yank, and the `lru`/`event-listener` unsoundness. The only
advisory that was ours to fix was `rustls-pemfile`, unmaintained and in the front
door's TLS path: it is gone, replaced by the `PemObject` trait on
`rustls-pki-types`, whose types `tls.rs` was already using on both sides of the
call.

The workspace manifest claimed member crates inherit shared versions "so versions
stay aligned repo-wide", and largely they did not: tokio was pinned separately in
eight crates, tracing in nine, anyhow in six, sha2 in four — none of them in the
table at all — while serde, serde_json and toml were bypassed in six more. Those
are centralized now, features still declared per crate and unioned by cargo. A
security bump to tokio is one edit rather than eight, and the eight can no longer
drift apart quietly.

Majors: sha2 0.11, toml 1, rand 0.10, rcgen 0.14, notify 8, tree-sitter 0.26,
tree-sitter-lua 0.5, syn 3. Four moved under the code. rcgen 0.14 signs against an
`Issuer` rather than a certificate plus a loose key, so the CA is a
`CertifiedIssuer` — the variant that still holds the certificate `ca.pem` is
written from. rand 0.10 renamed `OsRng` to `SysRng` and made it fallible, so the
salt and nonce paths state the panic they always documented. tree-sitter 0.26
indexes children by `u32` while still counting them as `usize`, which is 80 loops.

reqwest 0.12 and 0.13 were both being built; they now aren't. Worth knowing what
that cost, because it is not free: 0.13's `rustls` feature hard-wires aws-lc-rs,
which wants cmake and would put C crypto back into the static musl build, so this
takes `rustls-no-provider` — what iroh already selects, keeping the graph on ring.
The catch is that reqwest then installs no provider and *panics* when a client is
built. Installing it once at start-up is not enough and looked fine until the
tests said otherwise: they reach the request paths directly, never through `main`.
It is installed at each of the three client-construction sites instead.

Also latent, and not closed here: 0.13 drops webpki-roots entirely for the
platform verifier, so HTTPS trust now comes from the system store. On a minimal
Linux image with no /etc/ssl/certs that fails where 0.12 succeeded.
`use_preconfigured_tls()` is the way back if the node build needs it.

Two deps stay behind deliberately, documented at the pins. hf-hub 1.0 deletes
`ureq`, is reqwest-only, and its sole TLS feature is `reqwest/rustls` — the
aws-lc-rs one — which feature unification would then hand to every reqwest in the
workspace, breaking exactly the property apps/linux/build.sh documents. candle
0.11 uses the unstable `stdarch_neon_f16` intrinsics and does not compile on
stable aarch64 at all. tokenizers is pinned to 0.22 rather than latest, because
0.22 is what candle pulls and matching it is what removes the duplicate.

.github/workflows/deps.yml is the part that outlasts the version numbers. It
reads the lockfile, compiles nothing, and runs weekly as well as on push —
an advisory is not a property of a commit, and a push-only trigger means a quiet
week hides one. The two ignores are advisories with no fix available to us, listed
individually so a new unmaintained crate still fails the build.

+++ crates/adi-indexer/src/lang/cpp.rs
@@ -231,7 +231,7 @@ fn extract_cpp_symbols(node: Node, source: &str, symbols: &mut Vec<InternalSymbo
             }
         }
         _ => {
-            for i in 0..node.child_count() {
+            for i in 0..node.child_count() as u32 {
                 if let Some(child) = node.child(i) {
                     extract_cpp_symbols(child, source, symbols);
                 }
@@ -241,7 +241,7 @@ fn extract_cpp_symbols(node: Node, source: &str, symbols: &mut Vec<InternalSymbo
 }
 
 fn collect_class_members(body: Node, source: &str, children: &mut Vec<InternalSymbol>) {
-    for i in 0..body.child_count() {
+    for i in 0..body.child_count() as u32 {
         if let Some(child) = body.child(i) {
             match child.kind() {
                 "function_definition" | "declaration" => {
@@ -257,7 +257,7 @@ fn collect_class_members(body: Node, source: &str, children: &mut Vec<InternalSy
                         }
                 }
                 "field_declaration" => {
-                    for j in 0..child.child_count() {
+                    for j in 0..child.child_count() as u32 {
                         if let Some(field) = child.child(j)
                             && field.kind() == "field_identifier" {
                                 children.push(InternalSymbol {
@@ -364,7 +364,7 @@ fn collect_cpp_references(node: Node, source: &str, refs: &mut Vec<InternalRefer
         }
         "class_specifier" | "struct_specifier" => {
             if let Some(base) = node.child_by_field_name("base_clause") {
-                for i in 0..base.child_count() {
+                for i in 0..base.child_count() as u32 {
                     if let Some(child) = base.child(i)
                         && child.kind() == "base_class_clause"
                             && let Some(type_node) = child.child_by_field_name("type") {
@@ -379,7 +379,7 @@ fn collect_cpp_references(node: Node, source: &str, refs: &mut Vec<InternalRefer
         }
         _ => {}
     }
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_cpp_references(child, source, refs);
         }
