+++ crates/adi-indexer/src/lang/typescript.rs
@@ -38,7 +38,7 @@ fn extract_ts_symbols(node: Node, source: &str, symbols: &mut Vec<ParsedSymbol>)
                 let name_text = node_text(name, source);
                 let mut children = Vec::new();
                 if let Some(body) = node.child_by_field_name("body") {
-                    for i in 0..body.child_count() {
+                    for i in 0..body.child_count() as u32 {
                         if let Some(child) = body.child(i)
                             && child.kind() == "method_definition"
                                 && let Some(method_name) = child.child_by_field_name("name") {
@@ -120,7 +120,7 @@ fn extract_ts_symbols(node: Node, source: &str, symbols: &mut Vec<ParsedSymbol>)
             } else {
                 // Recurse into the export so e.g. `export class Foo`
                 // still emits a Class symbol.
-                for i in 0..node.child_count() {
+                for i in 0..node.child_count() as u32 {
                     if let Some(child) = node.child(i) {
                         extract_ts_symbols(child, source, symbols);
                     }
@@ -128,7 +128,7 @@ fn extract_ts_symbols(node: Node, source: &str, symbols: &mut Vec<ParsedSymbol>)
             }
         }
         _ => {
-            for i in 0..node.child_count() {
+            for i in 0..node.child_count() as u32 {
                 if let Some(child) = node.child(i) {
                     extract_ts_symbols(child, source, symbols);
                 }
@@ -214,7 +214,7 @@ fn collect_ts_references(node: Node, source: &str, refs: &mut Vec<ParsedReferenc
         }
         _ => {}
     }
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_ts_references(child, source, refs);
         }
