+++ crates/adi-indexer/src/parser/treesitter/analyzers/base.rs
@@ -96,7 +96,7 @@ impl AnalyzerBase {
         if ancestor.id() == target.id() {
             return true;
         }
-        for i in 0..ancestor.child_count() {
+        for i in 0..ancestor.child_count() as u32 {
             if let Some(child) = ancestor.child(i)
                 && Self::is_node_or_descendant_helper(child, target) {
                     return true;
